+++ cli/cli/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- `beam services ps` command has new `--federated` optional parameter which extends functionality with scanning for 
+IFederatedLogin and IFederatedInventory implementations among services in solution in current working directory 
+
 ## [1.16.2]
 
 ### Fixed